Add devDependency for npm's WebDriverjs
Please note that this is NOT the "official" one from Selenium:
cf. http://code.google.com/p/selenium/wiki/WebDriverJs vs.
    https://github.com/Camme/webdriverjs

The "official" one seems to not really be done yet. I compiled it
correctly but have not tested it. It requires the google Go! language
(at least for compiling) and much too much downloads. So I'll try
this thing from npm first, even if it's also not "complete" (the
author states so, see link above).